Last commit

+++ ChartRepoBackend/Models/DataSet.cs
@@ -3,6 +3,6 @@
     public class DataSet
     {
         public List<string> Labels { get; set; }
-        public List<decimal> Data { get; set; }
+        public List<int> Data { get; set; }
     }
 }
